requirements.txt
Django<=3.0.3
django-extensions<=2.2.8
django-filter<=2.2.0
djangorestframework<=3.11.0
mysqlclient<=1.4.6
oauthlib<=2.0.6
Pillow<=7.0.0
PyJWT<=1.5.3
PyMySQL<=0.9.3
sqlparse<=0.3.0
Twisted>=20.3.0
django-cors-headers<=3.2.1

